Rename security_sysinfo package and added a minor fix for windows.
