+++ internal/security_sysinfo/security_sysinfo_windows.go
@@ -1,7 +1,7 @@
 // Copyright 2023 New Relic Corporation. All rights reserved.
 // SPDX-License-Identifier: New Relic Pre-Release
 
-package sysinfo
+package security_sysinfo
 
 import (
 	"syscall"
